pipelines/requirements.txt
psycopg2
pandas
ruamel.yaml